refactor: 검증 관련 구현 코드 infra 패키지 하위로 이동
