+++ src/main/java/com/prography/lighton/member/users/infrastructure/validation/annotation/ValidPassword.java
@@ -1,10 +1,10 @@
-package com.prography.lighton.member.users.validation.annotation;
+package com.prography.lighton.member.users.infrastructure.validation.annotation;
 
 import static java.lang.annotation.ElementType.FIELD;
 import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
-import com.prography.lighton.member.users.validation.validator.ValidPhoneValidator;
+import com.prography.lighton.member.users.infrastructure.validation.validator.ValidPasswordValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 import java.lang.annotation.Retention;
@@ -12,12 +12,11 @@ import java.lang.annotation.Target;
 
 @Target({FIELD, PARAMETER})
 @Retention(RUNTIME)
-@Constraint(validatedBy = ValidPhoneValidator.class)
-public @interface ValidPhone {
-    String message() default "올바르지 않은 전화번호 형식입니다.";
+@Constraint(validatedBy = ValidPasswordValidator.class)
+public @interface ValidPassword {
+    String message() default "비밀번호는 영문, 숫자, 특수문자 포함 8자 이상이어야 합니다.";
 
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
 }
-
