+++ src/main/java/com/prography/lighton/member/users/infrastructure/validation/validator/ValidPasswordValidator.java
@@ -1,6 +1,6 @@
-package com.prography.lighton.member.users.validation.validator;
+package com.prography.lighton.member.users.infrastructure.validation.validator;
 
-import com.prography.lighton.member.users.validation.annotation.ValidPassword;
+import com.prography.lighton.member.users.infrastructure.validation.annotation.ValidPassword;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import java.util.regex.Pattern;
