+++ src/main/java/com/prography/lighton/member/users/presentation/dto/request/CompleteMemberProfileRequest.java
@@ -1,6 +1,6 @@
 package com.prography.lighton.member.users.presentation.dto.request;
 
-import com.prography.lighton.member.users.validation.annotation.ValidPhone;
+import com.prography.lighton.member.users.infrastructure.validation.annotation.ValidPhone;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
