+++ src/main/java/com/prography/lighton/member/users/presentation/dto/request/LoginMemberRequest.java
@@ -1,6 +1,6 @@
 package com.prography.lighton.member.users.presentation.dto.request;
 
-import com.prography.lighton.member.users.validation.annotation.ValidPassword;
+import com.prography.lighton.member.users.infrastructure.validation.annotation.ValidPassword;
 import jakarta.validation.constraints.Email;
 
 public record LoginMemberRequest(
